final/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.3
django-allauth==0.51.0
django-cache-url==3.4.2
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
environs==9.5.0
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
marshmallow==3.18.0
mock==4.0.3
mollie-api-python==2.13.0
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
psycopg2-binary==2.9.4
py==1.11.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sentry-sdk==1.10.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.12
